refactor: extract PageAnalyzerService
