+++ RagWebScraper/Services/IPageAnalyzerService.cs
@@ -1,0 +1,8 @@
+namespace RagWebScraper.Services;
+
+using RagWebScraper.Models;
+
+public interface IPageAnalyzerService
+{
+    Task<AnalysisResult?> AnalyzePageAsync(string url, List<string> keywords);
+}
